--- a/GPU/Common/GPUCommonDefAPI.h
+++ b/GPU/Common/GPUCommonDefAPI.h
@@ -133,7 +133,7 @@
     #define GPUglobal()
   #endif
   #define GPUconstant() __constant__
-  #define GPUconstexpr() __constant__
+  #define GPUconstexpr() constexpr __constant__
   #define GPUprivate()
   #define GPUgeneric()
   #define GPUbarrier() __syncthreads()