--- a/GPU/GPUTracking/Base/opencl-common/GPUReconstructionOCL.cl
+++ b/GPU/GPUTracking/Base/opencl-common/GPUReconstructionOCL.cl
@@ -78,7 +78,7 @@
 #define GPUCA_KRNL_LOAD_single(x_class, x_attributes, x_arguments, x_forward) GPUCA_KRNLGPU_SINGLE(x_class, x_attributes, x_arguments, x_forward)
 #define GPUCA_KRNL_LOAD_multi(x_class, x_attributes, x_arguments, x_forward) GPUCA_KRNLGPU_MULTI(x_class, x_attributes, x_arguments, x_forward)
 #define GPUCA_CONSMEM_PTR GPUglobal() char *gpu_mem, GPUconstant() MEM_CONSTANT(GPUConstantMem) * pConstant,
-#define GPUCA_CONSMEM *pConstant
+#define GPUCA_CONSMEM (*pConstant)
 #include "GPUReconstructionKernels.h"
 #undef GPUCA_KRNL
 #undef GPUCA_KRNL_LOAD_single