--- a/GPU/GPUTracking/Base/opencl-common/GPUReconstructionOCL.cxx
+++ b/GPU/GPUTracking/Base/opencl-common/GPUReconstructionOCL.cxx
@@ -25,8 +25,6 @@ using namespace GPUCA_NAMESPACE::gpu;
 #include <typeinfo>
 #include <cstdlib>
 
-constexpr size_t gGPUConstantMemBufferSize = sizeof(GPUConstantMem);
-
 #define quit(...)          \
   {                        \
     GPUError(__VA_ARGS__); \